Robot-Control-EEG/requirements.txt
scipy.io
pandas
numpy
matplotlib
tensorflow
seaborn
sklearn
keras